requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.215
langchainplus-sdk==0.0.17
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
pluggy==1.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.9
Pygments==2.15.1
pytest==7.4.0
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
SQLAlchemy==2.0.17
stack-data==0.6.2
tenacity==8.2.2
tiktoken==0.4.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
yarl==1.9.2
